+++ src/titiler/extensions/titiler/extensions/viewer.py
@@ -2,25 +2,24 @@
 
 from dataclasses import dataclass
 
+import jinja2
 from starlette.requests import Request
 from starlette.responses import HTMLResponse
 from starlette.templating import Jinja2Templates
 
 from titiler.core.factory import BaseTilerFactory, FactoryExtension
 
-try:
-    from importlib.resources import files as resources_files  # type: ignore
-except ImportError:
-    # Try backported to PY<39 `importlib_resources`.
-    from importlib_resources import files as resources_files  # type: ignore
-
-# TODO: mypy fails in python 3.9, we need to find a proper way to do this
-templates = Jinja2Templates(directory=str(resources_files(__package__) / "templates"))  # type: ignore
+DEFAULT_TEMPLATES = Jinja2Templates(
+    directory="",
+    loader=jinja2.ChoiceLoader([jinja2.PackageLoader(__package__, "templates")]),
+)  # type:ignore
 
 
 @dataclass
 class cogViewerExtension(FactoryExtension):
     """Add /viewer endpoint to the TilerFactory."""
+
+    templates: Jinja2Templates = DEFAULT_TEMPLATES
 
     def register(self, factory: BaseTilerFactory):
         """Register endpoint to the tiler factory."""
@@ -28,8 +27,8 @@ class cogViewerExtension(FactoryExtension):
         @factory.router.get("/viewer", response_class=HTMLResponse)
         def cog_viewer(request: Request):
             """COG Viewer."""
-            return templates.TemplateResponse(
-                name="cog_index.html",
+            return self.templates.TemplateResponse(
+                name="cog_viewer.html",
                 context={
                     "request": request,
                     "tilejson_endpoint": factory.url_for(request, "tilejson"),
@@ -44,14 +43,16 @@ class cogViewerExtension(FactoryExtension):
 class stacViewerExtension(FactoryExtension):
     """Add /viewer endpoint to the TilerFactory."""
 
+    templates: Jinja2Templates = DEFAULT_TEMPLATES
+
     def register(self, factory: BaseTilerFactory):
         """Register endpoint to the tiler factory."""
 
         @factory.router.get("/viewer", response_class=HTMLResponse)
         def stac_viewer(request: Request):
             """STAC Viewer."""
-            return templates.TemplateResponse(
-                name="stac_index.html",
+            return self.templates.TemplateResponse(
+                name="stac_viewer.html",
                 context={
                     "request": request,
                     "tilejson_endpoint": factory.url_for(request, "tilejson"),